+++ common/src/main/java/io/github/mortuusars/envelope/world/mail/dropoff/ServiceDropOffHandler.java
@@ -4,6 +4,8 @@ import com.mojang.logging.LogUtils;
 import io.github.mortuusars.envelope.api.ServiceDropOffHandlerRegistry;
 import io.github.mortuusars.envelope.world.item.component.mail.log.DeliveryRecord;
 import io.github.mortuusars.envelope.world.mail.address.type.ServiceAddress;
+import io.github.mortuusars.envelope.world.mail.service.EquineAssuranceBureau;
+import io.github.mortuusars.envelope.world.mail.service.ServiceAddresses;
 import net.minecraft.world.item.ItemStack;
 import org.slf4j.Logger;
 
@@ -35,6 +37,7 @@ public class ServiceDropOffHandler implements MailDropOffHandler {
 
         MailDropOffResult craftingResult = getCraftingHandler().handle(context);
         if (craftingResult.isHandled()) {
+            onCraftingHandled(context, address, craftingResult);
             return craftingResult;
         }
 
@@ -51,5 +54,14 @@ public class ServiceDropOffHandler implements MailDropOffHandler {
         }
 
         return MailDropOffResult.returned(mail, DeliveryRecord.Message.REJECTED);
+    }
+
+    // This is not clean code. But it's hard to find a place for logic like this in a data-driven system.
+    // Alternative would be to create a registry for callback types or something, which would be too much code.
+    // This will suffice for the time being.
+    protected void onCraftingHandled(MailDropOffContext context, ServiceAddress address, MailDropOffResult craftingResult) {
+        if (address.getDefinitionHolder().is(ServiceAddresses.EQUINE_ASSURANCE_BUREAU)) {
+            EquineAssuranceBureau.onCraft(context, address, craftingResult);
+        }
     }
 }